+++ pkg/resource/certmanager/generic_ca_issuer_builder.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
 	"fmt"
