+++ pkg/resource/certmanager/mtls_frontend_intermediate_ca_issuer_builder.go
@@ -15,39 +15,24 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
-	"fmt"
-
 	"github.com/alexandrevilain/temporal-operator/api/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-type WorkerFrontendClientCertificateBuilder struct {
-	GenericFrontendClientCertificateBuilder
+type MTLSFrontendIntermediateCAIssuerBuilder struct {
+	GenericCAIssuerBuilder
 }
 
-func NewWorkerFrontendClientCertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *WorkerFrontendClientCertificateBuilder {
-	return &WorkerFrontendClientCertificateBuilder{
-		GenericFrontendClientCertificateBuilder{
+func NewMTLSFrontendIntermediateCAIssuerBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSFrontendIntermediateCAIssuerBuilder {
+	return &MTLSFrontendIntermediateCAIssuerBuilder{
+		GenericCAIssuerBuilder: GenericCAIssuerBuilder{
 			instance:   instance,
 			scheme:     scheme,
-			name:       "worker-certificate",
-			secretName: "worker-certificate",
-			commonName: "Worker client certificate",
-			dnsName:    fmt.Sprintf("worker.%s", instance.ServerName()),
+			name:       "frontend-intermediate-ca-issuer",
+			secretName: "frontend-intermediate-ca-certificate",
 		},
 	}
-}
-
-func (b *WorkerFrontendClientCertificateBuilder) Update(object client.Object) error {
-	err := b.GenericFrontendClientCertificateBuilder.Update(object)
-	if err != nil {
-		return err
-	}
-
-	return controllerutil.SetControllerReference(b.instance, object, b.scheme)
 }
