+++ pkg/resource/certmanager/mtls_internode_intermediate_ca_certificate_builder.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
 	"fmt"
@@ -29,35 +29,35 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-type MTLSFrontendItermediateCACertificateBuilder struct {
+type MTLSInternodeItermediateCACertificateBuilder struct {
 	instance *v1alpha1.TemporalCluster
 	scheme   *runtime.Scheme
 }
 
-func NewMTLSFrontendIntermediateCACertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSFrontendItermediateCACertificateBuilder {
-	return &MTLSFrontendItermediateCACertificateBuilder{
+func NewMTLSInternodeIntermediateCACertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSInternodeItermediateCACertificateBuilder {
+	return &MTLSInternodeItermediateCACertificateBuilder{
 		instance: instance,
 		scheme:   scheme,
 	}
 }
 
-func (b *MTLSFrontendItermediateCACertificateBuilder) Build() (client.Object, error) {
+func (b *MTLSInternodeItermediateCACertificateBuilder) Build() (client.Object, error) {
 	return &certmanagerv1.Certificate{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      b.instance.ChildResourceName("frontend-intermediate-ca-certificate"),
+			Name:      b.instance.ChildResourceName("internode-intermediate-ca-certificate"),
 			Namespace: b.instance.Namespace,
 		},
 	}, nil
 }
 
-func (b *MTLSFrontendItermediateCACertificateBuilder) Update(object client.Object) error {
+func (b *MTLSInternodeItermediateCACertificateBuilder) Update(object client.Object) error {
 	certificate := object.(*certmanagerv1.Certificate)
 	certificate.Labels = object.GetLabels()
 	certificate.Annotations = object.GetAnnotations()
 	certificate.Spec = certmanagerv1.CertificateSpec{
 		IsCA:       true,
-		SecretName: b.instance.ChildResourceName("frontend-intermediate-ca-certificate"),
-		CommonName: "Frontend intermediate CA certificate",
+		SecretName: b.instance.ChildResourceName("internode-intermediate-ca-certificate"),
+		CommonName: "Internode intermediate CA certificate",
 		Duration:   b.instance.Spec.MTLS.CertificatesDuration.IntermediateCAsCertificates,
 		PrivateKey: &certmanagerv1.CertificatePrivateKey{
 			RotationPolicy: certmanagerv1.RotationPolicyAlways,
