+++ pkg/resource/certmanager/mtls_internode_intermediate_ca_issuer_builder.go
@@ -15,39 +15,24 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
-	"fmt"
-
 	"github.com/alexandrevilain/temporal-operator/api/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-type UIFrontendClientCertificateBuilder struct {
-	GenericFrontendClientCertificateBuilder
+type MTLSInternodeIntermediateCAIssuerBuilder struct {
+	GenericCAIssuerBuilder
 }
 
-func NewUIFrontendClientCertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *UIFrontendClientCertificateBuilder {
-	return &UIFrontendClientCertificateBuilder{
-		GenericFrontendClientCertificateBuilder{
+func NewMTLSInternodeIntermediateCAIssuerBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSInternodeIntermediateCAIssuerBuilder {
+	return &MTLSInternodeIntermediateCAIssuerBuilder{
+		GenericCAIssuerBuilder: GenericCAIssuerBuilder{
 			instance:   instance,
 			scheme:     scheme,
-			name:       "ui-mtls-certificate",
-			secretName: "ui-mtls-certificate",
-			commonName: "UI client certificate",
-			dnsName:    fmt.Sprintf("ui.%s", instance.ServerName()),
+			name:       "internode-intermediate-ca-issuer",
+			secretName: "internode-intermediate-ca-certificate",
 		},
 	}
-}
-
-func (b *UIFrontendClientCertificateBuilder) Update(object client.Object) error {
-	err := b.GenericFrontendClientCertificateBuilder.Update(object)
-	if err != nil {
-		return err
-	}
-
-	return controllerutil.SetControllerReference(b.instance, object, b.scheme)
 }
