+++ pkg/resource/certmanager/mtls_root_ca_certificate_builder.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
 	"fmt"
@@ -29,36 +29,36 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-type MTLSInternodeItermediateCACertificateBuilder struct {
+type MTLSRootCACertificateBuilder struct {
 	instance *v1alpha1.TemporalCluster
 	scheme   *runtime.Scheme
 }
 
-func NewMTLSInternodeIntermediateCACertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSInternodeItermediateCACertificateBuilder {
-	return &MTLSInternodeItermediateCACertificateBuilder{
+func NewMTLSRootCACertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSRootCACertificateBuilder {
+	return &MTLSRootCACertificateBuilder{
 		instance: instance,
 		scheme:   scheme,
 	}
 }
 
-func (b *MTLSInternodeItermediateCACertificateBuilder) Build() (client.Object, error) {
+func (b *MTLSRootCACertificateBuilder) Build() (client.Object, error) {
 	return &certmanagerv1.Certificate{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      b.instance.ChildResourceName("internode-intermediate-ca-certificate"),
+			Name:      b.instance.ChildResourceName("root-ca-certificate"),
 			Namespace: b.instance.Namespace,
 		},
 	}, nil
 }
 
-func (b *MTLSInternodeItermediateCACertificateBuilder) Update(object client.Object) error {
+func (b *MTLSRootCACertificateBuilder) Update(object client.Object) error {
 	certificate := object.(*certmanagerv1.Certificate)
 	certificate.Labels = object.GetLabels()
 	certificate.Annotations = object.GetAnnotations()
 	certificate.Spec = certmanagerv1.CertificateSpec{
 		IsCA:       true,
-		SecretName: b.instance.ChildResourceName("internode-intermediate-ca-certificate"),
-		CommonName: "Internode intermediate CA certificate",
-		Duration:   b.instance.Spec.MTLS.CertificatesDuration.IntermediateCAsCertificates,
+		Duration:   b.instance.Spec.MTLS.CertificatesDuration.RootCACertificate,
+		SecretName: b.instance.ChildResourceName("root-ca-certificate"),
+		CommonName: "Root CA certificate",
 		PrivateKey: &certmanagerv1.CertificatePrivateKey{
 			RotationPolicy: certmanagerv1.RotationPolicyAlways,
 			Encoding:       certmanagerv1.PKCS8,
@@ -69,7 +69,7 @@ func (b *MTLSInternodeItermediateCACertificateBuilder) Update(object client.Obje
 			b.instance.ServerName(),
 		},
 		IssuerRef: certmanagermeta.ObjectReference{
-			Name: b.instance.ChildResourceName("root-ca-issuer"),
+			Name: b.instance.ChildResourceName("bootstrap-issuer"),
 			Kind: certmanagerv1.IssuerKind,
 		},
 		Usages: []certmanagerv1.KeyUsage{
