+++ pkg/resource/certmanager/ui_frontend_client_certificate_builder.go
@@ -15,24 +15,39 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package resource
+package certmanager
 
 import (
+	"fmt"
+
 	"github.com/alexandrevilain/temporal-operator/api/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-type MTLSRootCAIssuerBuilder struct {
-	GenericCAIssuerBuilder
+type UIFrontendClientCertificateBuilder struct {
+	GenericFrontendClientCertificateBuilder
 }
 
-func NewMTLSRootCAIssuerBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *MTLSRootCAIssuerBuilder {
-	return &MTLSRootCAIssuerBuilder{
-		GenericCAIssuerBuilder: GenericCAIssuerBuilder{
+func NewUIFrontendClientCertificateBuilder(instance *v1alpha1.TemporalCluster, scheme *runtime.Scheme) *UIFrontendClientCertificateBuilder {
+	return &UIFrontendClientCertificateBuilder{
+		GenericFrontendClientCertificateBuilder{
 			instance:   instance,
 			scheme:     scheme,
-			name:       "root-ca-issuer",
-			secretName: "root-ca-certificate",
+			name:       "ui-mtls-certificate",
+			secretName: "ui-mtls-certificate",
+			commonName: "UI client certificate",
+			dnsName:    fmt.Sprintf("ui.%s", instance.ServerName()),
 		},
 	}
+}
+
+func (b *UIFrontendClientCertificateBuilder) Update(object client.Object) error {
+	err := b.GenericFrontendClientCertificateBuilder.Update(object)
+	if err != nil {
+		return err
+	}
+
+	return controllerutil.SetControllerReference(b.instance, object, b.scheme)
 }
